fix app extraction

+++ tests/ios/src/main/java/com/wearezeta/auto/ios/pages/IOSPage.java
@@ -240,13 +240,17 @@ public abstract class IOSPage extends BasePage {
         if (CommonUtils.getIsSimulatorFromConfig(this.getClass())) {
             IOSSimulatorHelper.goHome();
             Thread.sleep(timeSeconds * 1000);
-            final String ipaPath = (String) getDriver().getCapabilities().getCapability("app");
-            final File appPath = IOSCommonUtils.extractAppFromIpa(new File(ipaPath));
+            final String autPath = (String) getDriver().getCapabilities().getCapability("app");
             String bundleId;
-            try {
-                bundleId = IOSCommonUtils.getBundleId(new File(appPath.getCanonicalPath() + "/Info.plist"));
-            } finally {
-                FileUtils.deleteDirectory(appPath);
+            if (autPath.endsWith(".app")) {
+                bundleId = IOSCommonUtils.getBundleId(new File(autPath + "/Info.plist"));
+            } else {
+                final File appPath = IOSCommonUtils.extractAppFromIpa(new File(autPath));
+                try {
+                    bundleId = IOSCommonUtils.getBundleId(new File(appPath.getCanonicalPath() + "/Info.plist"));
+                } finally {
+                    FileUtils.deleteDirectory(appPath);
+                }
             }
             IOSSimulatorHelper.launchApp(bundleId);
             Thread.sleep(1000);
